Allow specifying a minimum number of files that must exist.
This allows preventing the removal of files in the following (hopefully rare)
cases:

  * The script creating backups has a failure and stops producing new entries
  * The system time errantly jumps forward (e.g. a typo on the date command)

This also has the following disadvantage:

  * If the backup cannot be made due to low disk space, old ones might not get
    purged to allow the backup to later succeed.

Of course, using minfiles thus becomes and administrative decision.

By default, this maintains the existing behavior (worst case, keeps a single
file to satisfy the oldest requested age) to prevent surprises.